% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_py_script.R
\name{check_py_script}
\alias{check_py_script}
\title{check_py_script}
\usage{
check_py_script(py_version)
}
\arguments{
\item{py_version}{Which version of python you want to use. One of '2' or '3'.}
}
\value{
Character vector
}
\description{
Checks that the py script to scrape tracks exists in your working directory
}
\examples{
check_py_script(2)
check_py_script("3")
}
