% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pages_api.R
\name{collect_tracks}
\alias{collect_tracks}
\title{collect_tracks}
\usage{
collect_tracks(username, api_key)
}
\arguments{
\item{username}{Your last.fm account username}

\item{api_key}{Your last.fm account api key}
}
\value{
A dataframe of songs and assoicated metadata
}
\description{
collect_tracks
}
