% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scrobbles.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(unix_col, to, ...)
}
\arguments{
\item{unix_col}{A column in a dataframe of 10 digit unix numbers}

\item{to}{One of 'Date' or 'Time'. Determines whether you get a datestamp of timestamp.}
}
\value{
Date vector
}
\description{
Convert a unix timestamp to either a datestamp or timestamp
}
\examples{
unix_time <- "1522124746"
timestamp <- convert(unix_time, to = "Time")

\dontrun{
# A dataframe read in from read_scrobbles called 'scrobbles'
scrobbles$Date <- convert(scrobbles$Date, to = "Time")
}
}
