% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tracks.R
\name{find_py_script}
\alias{find_py_script}
\title{find_py_script}
\usage{
find_py_script(version)
}
\arguments{
\item{version}{One of '2' or '3'. Specifies which version of python you will be running}
}
\value{
File path
}
\description{
Returns the file path to a py script
}
\examples{
find_py_script(version = "3")
}
