% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helpers.R
\name{get_last_timestamp}
\alias{get_last_timestamp}
\title{get_last_timestamp}
\usage{
get_last_timestamp(scrobbles_df, timestamp_column)
}
\arguments{
\item{scrobbles_df}{A dataframe output of `download_scrobbles`}

\item{timestamp_column}{The column in `scrobbles_df` that corresponds to the `date_unix` column}
}
\value{
Numeric vector
}
\description{
get_last_timestamp
}
\examples{
\dontrun{
get_last_timestamp(mydata, 'date_unix')
}
}
