% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_lastfm_credentials}
\alias{get_lastfm_credentials}
\title{get_lastfm_credentials}
\usage{
get_lastfm_credentials(env)
}
\arguments{
\item{env}{One of 'username' or 'key', to return the username or key environment variable respectively}
}
\value{
Character string
}
\description{
get_lastfm_credentials
}
\examples{
\dontrun{
get_lastfm_credentials(env = 'username')
get_lastfm_credentials(env = 'key')
}
}
