% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_py_script.R
\name{install_scrobble_script}
\alias{install_scrobble_script}
\title{install_export_script}
\usage{
install_scrobble_script(version)
}
\arguments{
\item{version}{Which version of python to run. One of '2' or '3'.}
}
\value{
Invisibly returns TRUE or FALSE, indicating whether the file copy was successful.
}
\description{
Copies the script used to scrape scrobbles into your working directory. Necessary in order
to run 'fetch_tracks'
}
\examples{
install_scrobble_script(version = "3")
}
