% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scrobbles.R
\name{read_scrobbles}
\alias{read_scrobbles}
\title{read_scrobbles}
\usage{
read_scrobbles(file, convert_time = "None", ...)
}
\arguments{
\item{file}{A .txt file from the output of `fetch_tracks`}

\item{convert_time}{One of 'None', 'Date', 'Time'. Determines what format to put the Date
column in. Either as a datestamp ('Date'), timestamp ('Time'), or (the default) left as a
UNIX timestamp ('None').}

\item{...}{Additional arguments to pass to `read_delim`}
}
\value{
A dataframe
}
\description{
Wrapper around read_delim to correctly read a scrobbled tracks file.
}
\examples{
my_tracks <- read_scrobbles(system.file("extdata", "scrobbles.txt", package = "scrobbler"),
convert_time = "Date")
}
