% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrobbler.R
\name{setup_scrobbler}
\alias{setup_scrobbler}
\title{setup_scrobbler}
\usage{
setup_scrobbler()
}
\description{
Place python file fetching scrobbled tracks into your working directory
}
\examples{
\dontrun{
library(musicr)
setup_scrobbler()
}
}
