% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pages_api.R
\name{download_scrobbles}
\alias{download_scrobbles}
\title{download_scrobbles}
\usage{
download_scrobbles(
  username = get_lastfm_credentials("username"),
  api_key = get_lastfm_credentials("key")
)
}
\arguments{
\item{username}{Your last.fm account username}

\item{api_key}{Your last.fm account api key}
}
\value{
A dataframe of songs and associated metadata
}
\description{
download_scrobbles
}
\examples{
\dontrun{
download_scrobbles(username = "your_username", api_key = "your_api_key")
}
}
