% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pages_api.R
\name{download_scrobbles}
\alias{download_scrobbles}
\title{download_scrobbles}
\usage{
download_scrobbles(username, api_key)
}
\arguments{
\item{username}{Your last.fm account username}

\item{api_key}{Your last.fm account api key}
}
\value{
A dataframe of songs and assoicated metadata
}
\description{
download_scrobbles
}
\examples{
\dontrun{
download_scrobbles(username = "your_username", api_key = "your_api_key")
}
}
