% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tracks.R
\name{fetch_tracks}
\alias{fetch_tracks}
\title{fetch_tracks}
\usage{
fetch_tracks(username, out_file, start_page = NULL)
}
\arguments{
\item{username}{Username of last.fm account to fetch scrobbles from}

\item{out_file}{Name of output file to save tracks to (i.e., scrobbles.txt)}

\item{start_page}{Page to start from. Defaults to 1}
}
\value{
.txt file of scrobbled tracks
}
\description{
Runs a script in your working directory called 'lastexport.py'
}
\examples{
\dontrun{
fetch_tracks("your_username", out_file = "scrobbles.txt", start_page = 1)
}
}
