% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pages_api.R
\name{update_scrobbles}
\alias{update_scrobbles}
\title{update_scrobbles}
\usage{
update_scrobbles(
  data,
  timestamp_column,
  username = get_lastfm_credentials("username"),
  api_key = get_lastfm_credentials("key")
)
}
\arguments{
\item{data}{A dataframe outputted by `download_scrobbles`}

\item{timestamp_column}{The `date_unix` column in your dataframe}

\item{username}{Last.fm API username}

\item{api_key}{Last.fm API key}
}
\value{
A dataframe
}
\description{
Companion function to `download_scrobbles`. Only downloads the scrobbles that have been
stored since you ran `download_scrobbles`.
}
\examples{
\dontrun{
mydat <- download_scrobbles(username = "your_username", api_key = "your_api_key")
update_dat <- update_scrobbles(mydat,
    "date_unix",
    username = "your_username",
    api_key = "your_api_key")
}
}
